colorpick

+++ coolor/events/enums.go
@@ -4,12 +4,12 @@ import "strconv"
 
 // EnumName associates an enum value with its name for printing.
 type EnumName struct {
-	V uint32
+	V uint64
 	S string
 }
 
 // EnumString converts a flag-based enum value into its string representation.
-func EnumString(v uint32, names []EnumName, goSyntax bool) string {
+func EnumString(v uint64, names []EnumName, goSyntax bool) string {
 	s := ""
 	for _, n := range names {
 		if v&n.V == n.V && (n.V != 0 || v == 0) {
